fixup and reorganize
