+++ package.json
@@ -10,9 +10,12 @@
     }
   },
   "types": "./types/index.d.ts",
+  "files": [
+    "/lib",
+    "/types"
+  ],
   "engines": {
-    "node": ">=16.0.0",
-    "npm": ">=8.1.4"
+    "node": ">=20.11.0"
   },
   "scripts": {
     "lint": "eslint \"./lib/**/*.{js,mjs}\" \"./types/**/*.d.ts\"",
@@ -153,13 +156,13 @@
     "@types/node": "^18.7.23",
     "@typescript-eslint/eslint-plugin": "^5.38.1",
     "@typescript-eslint/parser": "^5.38.1",
-    "@xan105/nodert": "^1.1.0",
+    "@xan105/nodert": "^1.1.16",
     "eslint": "^8.22.0",
     "tap": "^16.3.0",
     "typescript": "^4.7.4"
   },
   "dependencies": {
-    "@xan105/error": "^1.4.3",
+    "@xan105/error": "^1.7.0",
     "@xan105/fs": "^2.1.0",
     "@xan105/is": "^2.7.0"
   }
